Rename the DOM manipulation library from El to DOM. Rename the Three scene graph manipulation library from Graph to SOM (Spatial Object Model).
